% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{abcAS}
\alias{abcAS}
\title{Wrapper to apply the ABC parameter inference}
\usage{
abcAS(
  population,
  ploidy,
  refTable,
  sumStats = c("lambda", "rbarD", "Pareto"),
  ...
)
}
\arguments{
\item{population}{The population to infer the sexual reproduction rate. This can be the path to a Genalex formatted .csv file, or an object of the classes genind, genclone, structure, genetix or genpop (see \link[poppr]{poppr} for details)}

\item{ploidy}{Integer. The ploidy level of the population. Required **only** when a path to a Genalex .csv file is used.}

\item{refTable}{data.frame. The data frame containing the reference table of the simulated populations.}

\item{sumStats}{Vector. The names of the summary statistics to be calculated of the population.}

\item{...}{Additional parameters to be passed to \link[abc]{abc}.}
}
\value{
Returns an object of class \link[abc]{abc}.
}
\description{
This function is a wrapper to facilitate the use inference of the sexual reproduction rate using approximate Bayesian Computation implemented in the package `abc`. For more details see \link[abc]{abc}
}
