% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{addinds}
\alias{addinds}
\title{Create "indiv" objects from a matrix of microsatellite data.}
\usage{
addinds(x, mrows)
}
\arguments{
\item{x}{Matrix with the microsatellite repetition data.}

\item{mrows}{Ploidy level of the population.}
}
\description{
This function takes a matrix of microsatellite loci and creates an "indiv" object with this information.
}
