% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{asexrep}
\alias{asexrep}
\title{Generate copies of the objects of class "indiv"}
\usage{
asexrep(x)
}
\arguments{
\item{x}{Individual to be replicated}
}
\description{
Used to simulate asexual/clonal reproduction in a population of individuals of the class "indiv"
}
