% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{Beta}
\alias{Beta}
\title{Function to apply the calculation of the negative slope of the Pareto distribution as described in the \code{poppr} package.}
\usage{
Beta(x)
}
\arguments{
\item{x}{Abundance of a particular genotype.}
}
\value{
Negative slope of the Pareto distribution from a genotype table.
}
\description{
This function is a wrapper to apply the calculation of the negative slope of the Pareto distribution to a table of genotypes.
}
