% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{cv4abcAS}
\alias{cv4abcAS}
\title{Wrapper to apply the cross validation for ABC}
\usage{
cv4abcAS(refTable, sumStats = c("lambda", "rbarD", "Pareto"), ...)
}
\arguments{
\item{refTable}{`data.frame`. The reference table from the simulated populations.}

\item{sumStats}{Vector. Selected summary statistics.}

\item{...}{Additional parameters to be passed to \link[abc]{cv4abc}.}
}
\value{
An object of class \link[abc]{cv4abc}
}
\description{
This function is a wrapper to facilitate the use of the cross validation for approximate Bayesian Computation using the package `abc`. For more details see \link[abc]{cv4abc}
}
