% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{displacement}
\alias{displacement}
\title{Function to sample the new location of a dispersed individual}
\usage{
displacement(x, position, probability)
}
\arguments{
\item{x}{Individual of the class "indiv" with a position slot.}

\item{position}{Spatial matrix.}

\item{probability}{Dispersal kernel. Array of transition probabilities for the movement from each possible slot.}
}
\value{
Sampled new position within the position matrix given as a parameter.
}
\description{
This function samples the new position of an individual in the position matrix given the current position of the individual and the dispersal kernel.
}
