% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{distint}
\alias{distint}
\title{Calculate the probability of dispersal to each slot}
\usage{
distint(i, j, mat)
}
\arguments{
\item{i}{Initial quadrant}

\item{j}{Arrival quadrant}

\item{mat}{Position matrix where the individuals are located}
}
\value{
The probability of dispersal from a site i to all possible quadrants of the defined position grid
}
\description{
This function uses the eucprob function to calculate the probability density contained in each slot depending on each origin site.
}
