% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{distmat}
\alias{distmat}
\title{Calculate the probability of dispersal from each initial quadrant to each possible quadrant}
\usage{
distmat(x, mat)
}
\arguments{
\item{x}{The quadrant of origin.}

\item{mat}{The probability matrix of the displacement rates.}
}
\value{
It returns an array of probabilities of displacement from each of the possible origin quadrants.
}
\description{
This function calculates the probabilities of displacement from an x quadrant to every other quadrant in the grid.
}
