% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{evoSim_se}
\alias{evoSim_se}
\title{Spatially explicit evolutionary simulation of organisms with mixed reproduction}
\usage{
evoSim_se(
  n = 100,
  msat = 10,
  gen = 100,
  ploidy = 1,
  sexprob = 0.5,
  mutrat = 0.001,
  grid = c(10, 10),
  printAnc = FALSE
)
}
\arguments{
\item{n}{Integer. Number of individuals samples for reproduction each generation. This is the number of individuals that will be considered for reproduction. This is done to reduce the computational burden by having an infinitely exponentially growing population. This will also be a proxy of the fact that sampled populations are a small sample of all the individuals that compose an actual population.}

\item{msat}{Double. Amount of microsatellite loci for your individuals.}

\item{gen}{Integer. Amount of generations to pass before returning the resulting population.}

\item{ploidy}{Integer. Ploidy level of the individuals in the population.}

\item{sexprob}{Double. Conditional probability of sexual reproduction events.}

\item{mutrat}{Double. Mutation rate of the microsatellite loci.}

\item{grid}{Integer vector. Dimensions of the spatial grid where the populations occurs.}

\item{printAnc}{Boolean. If \code{TRUE} the genotype of the first individual created will be printed at the end of the simulation.}
}
\value{
A named list with two slots. The \code{pop} slot that contains the Genind object containing the simulated population, and the \code{sexprop} slot that contains the proportion of sexual reproduction events in each generation.
}
\description{
This function performs a spatially explicit evolutionary simulation of a population with mixed reproduction.
}
\details{
This function allows to specify the dimensions of the position matrix, the probability of sexual reproduction, the amount of microsatellite markers, the ploidy, the amount of generations and the mutation rate.
   This function starts by creating an individual starting in a random position of the position grid. This individual will start the first generation by reproducing clonaly. The individuals generated by clonal reproduction will disperse to other quadrants in the grid given the dispersal kernel. Once there are multiple individuals, the individuals that are located in the same quadrant are able to reproduce sexually. The sexual reproduction probability is handled as a conditional probability. The sexual reproduction then has a probability of \code{sexprob} given that the individuals of the crossing are present in the same quadrant (representing here the same site).
   After generating the individuals of the next generation, the mutations are introduced to the microsatellite loci with a binomial probability distribution given by p = \code{mutrat}. This will simulate the accumulation of mutations along generations of reproduction of the organisms.
}
\examples{
# Simulate a testing population.
population <- evoSim_se(n=10, msat=10, gen=100, ploidy=3, sexprob=0.2, mutrat=0.001, grid=c(3,3))
}
