% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\docType{class}
\name{indiv-class}
\alias{indiv-class}
\title{Class individual for evolutionary simulations}
\description{
Class that contains a matrix of the microsatellite alleles for multiple loci and multiple ploidy levels. It also includes a position slot for spatially explicit inferences.
}
\section{Fields}{

\describe{
\item{\code{micro}}{Matrix that contains the microsatellite alleles (as rows) for each locus (as columns).}

\item{\code{pos}}{Integer. Indicator of the position to facilitate spatially explicit models.}
}}

