% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{insMutations}
\alias{insMutations}
\title{Insert mutations in microsatellite markers following a step wise mutation model}
\usage{
insMutations(x, mutationRate)
}
\arguments{
\item{x}{"indiv" individual where the mutations will be introduced.}

\item{mutationRate}{Mutation rate of the microsatellite markers.}
}
\description{
Inserts a step wise mutation in a microsatellite allele using a binomial probability distribution with the probability specified as the mutationRate parameter.
}
