% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{makeRefTable}
\alias{makeRefTable}
\title{Function to make the reference table from a pool of simulated populations.}
\usage{
makeRefTable(
  N,
  pop_number = 100,
  sample_size = NULL,
  generations = 10000,
  ploidy = 1,
  loci = 10,
  mutation_r = 0.001,
  file_name = "reftable.csv",
  spgrid = c(10, 10),
  cores = NULL
)
}
\arguments{
\item{N}{Integer. Number of populations to be simulated.}

\item{pop_number}{Integer. Limit number of individuals to reproduce per generation per population.}

\item{sample_size}{Integer. Passed on to the \code{evoSim_se} function. Individuals to be sampled to calculate the summary statistics. If \code{sample_size} is more than \code{pop_number}, all individuals will be used for the calculations.}

\item{generations}{Integer. Passed on to the \code{evoSim_se} function. Number of generations simulate for each population.}

\item{ploidy}{Integer. Passed on to the \code{evoSim_se} function. Ploidy level of the individuals in the simulated populations.}

\item{loci}{Integer. Passed on to the \code{evoSim_se} function. Number of microsatellite loci to be simulated for each individual in the population.}

\item{mutation_r}{Double. Passed on to the \code{evoSim_se} function. Mutation rate of the microsatellite markers.}

\item{file_name}{String. Name of the file to be created to store the reference table.}

\item{spgrid}{Int vector. Passed on to the \code{evoSim_se} function. The dimensions of the spatial grid where the populations are simulated.}

\item{cores}{Integer. Amount of parallel cores to run the simulations. **Warning** Windows does not support multi-core execution, so to run this function in Windows the `cores` parameter should be set to 1.}
}
\value{
This function does not return a value. It writes out the reference table using the name given in `file_name` to the working directory.
}
\description{
This function provides an easy-to-use wrapper to generate a pool of simulated populations and calculate the summary statistics from each simulated population.
The function samples \code{N} values from a uniform probability distribution between 0 and 1. Each one of the sampled values will be used as a probability of sexual reproduction for each of the simulated populations.
The summary statistics are collected in a table that is written into a file at the working directory.
Because of the computationally intensive nature of the procedure this function can be executed in several different runs. If the file exists already the function adds a number to the file name to preserve previous runs.
}
