% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{mutmsat}
\alias{mutmsat}
\title{Introduce step wise mutations to a microsatellite marker}
\usage{
mutmsat(y, mutrate)
}
\arguments{
\item{y}{Integer. Number of repetitions of the allele.}

\item{mutrate}{Double. This is the mutation rate for the locus.}
}
\description{
This function is used to introduce mutations to microsatellite markers following a stepwise mutation model. Following this, the mutated allele will be either a repetition smaller or a repetition bigger than the wild type allele.
}
