% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{pastealleles}
\alias{pastealleles}
\title{Paste the alleles of an individual separated by a semicolon}
\usage{
pastealleles(x)
}
\arguments{
\item{x}{Vector of microsatellite loci to be collapsed.}
}
\description{
This function collapses the alleles separating them with a semicolon.
}
