% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description.R
\docType{data}
\name{Pinf_China}
\alias{Pinf_China}
\title{Microsatellite genotypes of the Colombian population}
\format{
Object of the clas Genclone as defined by the adegenet R package.
}
\source{
https://onlinelibrary.wiley.com/doi/10.1111/jeu.12010
}
\usage{
Pinf_China
}
\description{
A set of samples of *P. infestans* collected in Korea and characterized by 8 microsatellite loci: Pi4B Pi02 G11 Pi33 Pi56 Pi16 Pi70 Pi63
}
\keyword{datasets}
