% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description.R
\docType{data}
\name{Pinf_Korea}
\alias{Pinf_Korea}
\title{Microsatellite genotypes of the Colombian population}
\format{
Object of the clas Genclone as defined by the adegenet R package.
}
\source{
https://bsppjournals.onlinelibrary.wiley.com/doi/10.1111/ppa.13178
}
\usage{
Pinf_Korea
}
\description{
A set of samples of *P. infestans* collected in Korea and characterized by 12 microsatellite loci: D13, G11, Pi02, Pi04, Pi4B, Pi63, Pi70, SSR11, SSR2, SSR4, SSR6, and SSR8.
}
\keyword{datasets}
