% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description.R
\docType{data}
\name{Pinf_Mexico}
\alias{Pinf_Mexico}
\title{Microsatellite genotypes of the Colombian population}
\format{
Object of the clas Genclone as defined by the adegenet R package.
}
\source{
https://doi.org/10.1111/mec.14000
}
\usage{
Pinf_Mexico
}
\description{
A set of samples of *P. infestans* collected in Mexico and characterized by 12 microsatellite loci: D13, G11, Pi02, Pi04, Pi4B, Pi63, Pi70, SSR11, SSR2, SSR4, SSR6, and SSR8.
}
\keyword{datasets}
