% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{posRecomb}
\alias{posRecomb}
\title{Obtains the position of a descendant individual from a sexual crossing}
\usage{
posRecomb(x, elList)
}
\arguments{
\item{x}{Parental individual.}

\item{elList}{Population list where all the individuals are stored.}
}
\value{
It returns the position of the parent x.
}
\description{
After a sexual recombination, this function is used to obtain the location of the parental that is the place where the descendant individual appears.
}
