% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{power_law_beta}
\alias{power_law_beta}
\title{Pareto distribution using \code{poweRlaw}}
\usage{
power_law_beta(x)
}
\arguments{
\item{x}{Value used to generate the Pareto distribution}
}
\value{
Returns the negative slope of the Pareto distribution
}
\description{
This function calcultates the negative slope of a Pareto distribution using the \code{poweRlaw} package as described in the \code{poppr} package.
}
