% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{recombination}
\alias{recombination}
\title{Function to do the recombination of the alleles in a locus}
\usage{
recombination(x, size, poplist)
}
\arguments{
\item{x}{"indiv" individual index.}

\item{size}{Ploidy level of the individuals.}

\item{poplist}{Population list. List of the "indiv" individuals.}
}
\description{
Selects alleles from an individual locus for a n-ploid individual.
}
