% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{refTab}
\alias{refTab}
\title{Function used to generate a reference table from multiple simulations.}
\usage{
refTab(
  x,
  pop_number,
  generations,
  ploidy,
  loci,
  mutation_r,
  spgrid,
  sample_size
)
}
\arguments{
\item{x}{Element of a vector or list of values for the probability of sexual reproduction.}
}
\value{
List of the given probability of sexual reproduction, the proportion of sexual reproduction events, and the summary statistics for each simulated population.
}
\description{
This function is a wrapper to implement the simulation and summarizing of the populations pooled for the ABC inference of the proportion of sexual reproduction.
This allows to take advantage of the multi-core execution available from the \code{parallel} R package to make the production of a reference table more efficient.
}
