% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{refTabComp}
\alias{refTabComp}
\title{Function to reunite the different parts of the reference table obtained in different runs}
\usage{
refTabComp(fpath, folder = TRUE)
}
\arguments{
\item{fpath}{String. Name of the folder that contains the parts of the reference table. If `folder` is `FALSE` then is the pattern (or common string) of the names of the parts of the reference table.}

\item{folder}{Boolean. Default `TRUE`. If `TRUE` the function will combine all parts in the folder specified in `fpath`. If `FALSE`, all files whose name contain `fpath` are combined into the reference table. Used in case the reference table parts are in the same working directory.}
}
\value{
Reference table containing the parameter values and the summary statistics of all the simulations done in different runs.
}
\description{
Due to the computation intensive nature of the simulations required for the ABC inference, the simulations might be done in several different runs. This function allows to easily combine the parts of a reference table for an inference obtained in different runs.
}
