% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFns.R
\name{sim2genind}
\alias{sim2genind}
\title{Converts a simulated population into a Genind object}
\usage{
sim2genind(x, ploidy)
}
\arguments{
\item{x}{List of objects of the class "indiv" produced during the evolutionary simulation.}
}
\value{
Genind of the individuals of the list produced during the simulation.
}
\description{
This function converts a simulated population into a Genind object. This in order to facilitate the calculation of diversity statistics using the poppr package.
}
