% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_inference_tools.R
\name{summaryStats}
\alias{summaryStats}
\title{Function to calculate the summary statistics from the population}
\usage{
summaryStats(population, ploidy, sumStats = c("lambda", "rbarD", "Pareto"))
}
\arguments{
\item{population}{The population used to calculate the summary statistics. This can be the path to a Genalex formatted .csv file, or an object of the classes genind, genclone, structure, genetix or genpop (see \link[poppr]{poppr} for details)}

\item{ploidy}{Integer. The ploidy level of the population. Required only when a path to a Genalex .csv file is used.}

\item{sumStats}{Vector. The names of the summary statistics to be calculated of the population.}
}
\value{
Named vector containing the values for the summary statistics specified in the function call.
}
\description{
This function calculates the selected summary statistics from a population. Warning because the ABC inference of the sexual reproduction rate can be done on one population at a time, the sub-population strata will be ignored and the population given will be treated as a single population. The available summary statistics are:

The number of multi-loci genotypes `MLG`,
the expected number of MLGs `eMLG`, the Shannon-Weiner Diversity index `H`,
the Stoddard and Taylor's Index `G`,
the Simpson's index `lambda`,
the evenness `E.5`,
Nei's gene diversity (expected heterozygosity) `Hexp`,
a numeric vector giving the value of the Index of Association for each population factor `Ia`,
a numeric vector giving the value of the Standardized Index of Association for each population factor `rbarD`,
as implemented by the \link[poppr]{poppr};
and the beta Pareto coefficient for the MLG distribution `Pareto`.
}
